+++ NEWS.md
@@ -20,6 +20,8 @@ This is an update that corresponds to Python Polars 1.38.1.
   ([pola-rs/polars#26437](https://github.com/pola-rs/polars/pull/26437)).
 - New `<expr>$bin$get()` to extract a specific byte from a binary value (#1731).
 - `<expr>$str$split()` has two new arguments `literal` and `strict` (#1730).
+- New `pl$scan_lines()` and `pl$read_lines()` to read one or several files into
+  a single column (#1732).
 
 ### Bug fixes